refactor field retrieval
