+++ code/system.ui/react.ui.monaco/src/ui.logic/MonacoCrdt/-dev/-TEST.mts
@@ -4,16 +4,4 @@ export default Dev.describe('MonacoCrdt', (e) => {
   e.it('tmp', async (e) => {
     expect(123).to.eql(123);
   });
-  e.it('tmp', async (e) => {
-    expect(123).to.eql(123);
-  });
-  e.it('tmp', async (e) => {
-    expect(123).to.eql(123);
-  });
-  e.it('tmp', async (e) => {
-    expect(123).to.eql(123);
-  });
-  e.it('tmp', async (e) => {
-    expect(123).to.eql(123);
-  });
 });
